requirements.txt
requests
python-dotenv
selenium
webdriver-manager
beautifulsoup4
fastapi
apscheduler
uvicorn
jinja2